jz4740: PWM: Fix compilation, invert duty cycle, fix duty=0 behaviour, switch to "ext" clock source.
The active part of the cycle is after the "half" value has been reached, so the "half" value should be smaller for longer activity.

The hardware behaviour when the "half" value is 0 depends on the input clock frequency (as determined source and prescaler). Experiments have shown that the EXT clock with prescale 1 works.

Another advantage of the EXT clock is that its frequency is constant, while the PCLK frequency is derived from the PLL and my vary at runtime, invalidating our computed "half" and "full" values.