a320_defconfig: Disable serial port and console logging over it.
Logging over serial port was slowing down the boot a lot.
Those few people who have a serial port on their Dingoo are perfectly capable of building their own kernel with support for it.
In fact, probably the only reason to add a serial port is to test your own kernels.