a320: Use 8 bits per block in bad block table.
We found a Dingoo with at least one bad block on it and it showed 1 sequence of 8 successive bad blocks, starting at a multiple of 8.
I am assuming it is far more likely that 8 bits per block are used in the bad block table than that exactly those 8 blocks would be bad and no other blocks.