MIPS: JZ4740: SLCD framebufer driver.
This driver sends the frame buffer to a smart LCD controller, that is a controller with its own video memory.
It is a squashed version of development done in the jz-2.6.34 branch.
I changed one thing compared to that branch: the platform resource and device are now shared with the normal LCD controller, since they actually share the same hardware resources and therefore only one of the two drivers can be used.