MIPS: JZ4740: cpufreq: Fix section mismatch warning.
The cpufreq_jz4740_driver variable containg a function pointer to jz4740_cpufreq_driver_init which was marked __init. I removed that __init.

An alternative would be to mark cpufreq_jz4740_driver using __ref: this does silence the warning, but none of the other cpufreq drivers I checked use this approach.
Also, I could not find any documentation about when the init pointer from the struct is used exactly, so it is also safer not to consider jz4740_cpufreq_driver_init an init routine.