MIPS: JZ4740: A320: Removed part of command line that disables DMA in USB gadget driver.
The DMA was disabled by booboo to work around a bug where the USB connection would get stuck once in a while.
This bug may no longer be present in the latest kernel; I am removing the workaround so we can test this.
In a quick test I did not find any problems.
Also it seems that Qi Hardware are not using the workaround.