MIPS: A320: defconfig: Disable OSS compatibility drivers.
We need software volume because the JZ4740 hardware only supports 4 levels, all of which are loud.
ALSA implements software volume in libalsa, while the OSS compatibility drivers send the audio directly to the ALSA drivers, thus bypassing libalsa and not applying the software volume.
In addition, it is easier to debug sound problems if we know which audio path is used, since there is only one possibility.
For applications that are dynamically linked to SDL, the migration to ALSA should be transparent.