MIPS: JZ4740: A320: Force backlight off when ILI is off.
When TV-out is active, the controller for the internal screen (ILI9325/9331) is off.
This patch ensures that the backlight is turned off for as long as the ILI chip stays off,
even if the user changes the brightness value via sysfs in the mean time.

Thanks to Paul Cercueil <paul@crapouillou.net> for the idea of reading the ILI RESET_N pin
to determine whether TV-out is active.