at86rf230: separate state change command from final state
at86rf230_state expects the number of the final state to be equal to
the number of the state change command. This is not true for commands
that force a state change, such as STATE_FORCE_TRX_OFF and
STATE_FORCE_TX_ON.

at86rf230_state would thus erroneously return an error, which caused
at86rf230_xmit to exit before sending the frame.

This patch separates the state change command passed to at86rf230_state
from the final state.

- at86rf230.c (at86rf230_state): separate state change command and final
  state, since not all commands lead to a state with the same number
- at86rf230.c (at86rf230_start, at86rf230_stop, at86rf230_xmit): updated
  for new use of at86rf230_state

Signed-off-by: Werner Almesberger <werner@almesberger.net>