MTD: NAND: JZ4740: Finished multi-bank support and added autodetection
The platform data can now specify which external memory banks to probe
for NAND chips, and in which order. Banks that contain a NAND are used
and the other banks are freed.

Thanks to Paul Cercueil <paul@crapouillou.net> for the initial patch that
performed autodetection.

Additionally, the following problems were tackled:
- add_mtd_device returns 0/1 instead of an error code
- the mmio resource should be released if probe fails
- avoid freeing non-existing busy gpio