OpenDingux: initrd: Specify initrd contents using text file
The gen_init_cpio tool will parse the text file and generate a cpio archive.

splashkill is a simple program that is meant to be launched directly by the
kernel, before mininit. It forks itself and sleeps until the ENTER key is
pressed; when it occurs, it enables the framebuffer console by writting to
a file in /sys. This permits to 'kill' the kernel splash screen, in order
to display the boot log.

The source for the mininit and splashkill binaries can be found here:
  https://github.com/Ayla-/mininit