OpenDingux: initrd: Specify initrd contents using text file instead of cpio archive
The gen_init_cpio tool will parse the text file and generate a cpio archive
after all, but the files in git are more accessible now.

I put in all device nodes and directories that were present in the cpio archive
that we were using. Probably not all of it is actually needed.