MIPS: JZ4740: A320: defconfig: Disable CPU frequency scaling for now.
I get I/O errors reading from the SD card when the cpufreq driver is
used. In earlier kernels we had SD problems with cpufreq and suspend,
but in 2.6.39 it also shows up under normal use. Until this problem
is fixed in the cpufreq driver, it's better to disable it.