MIPS: JZ4740: cpufreq: Set CE bit after PLL freq change.
This fixes I/O errors when reading from SD card.
I guess this is a hardware quirk: I even tried writing the current value
with the same routine and that had no effect, so it is not some side effect
that avoids the I/O errors, it is actually the CE bit that matters.