spi/atusb: Initial setup to be a fake spi master driver.
Parts of it are based on the spi_atben driver. We have USB on one side and
fakinf SPI on the other side here. We then "connnect" the already working
at86rf230 driver to this spi master for the actual chip communication.

Not working yet but at least the driver register.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>