spi/atusb: Big update and sync with atben driver.
I hacked to much stuff together and was to lazy to split it all of in tiny
pieces. My bad.

The at86rf230 driver now gets attached to atusb and tries to request data from
us during probe. This fails as we do sync usb and the spi layer holds a spinlock
so we schedule while atomic. Re-work to async USB is the next step.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>