atusb: added tentative interrupt support
Added the USB-to-interrupt logic. The approach is basically to have a
bulk URB that waits for an incoming interrupt notification and that
gets resubmitted whenever the interrupt is acknowledged.

We map enable/disable to an atomic count of the masking depth.

To do:
- convert printks to dev_dbg or remove them
- remove bulk URB on driver unload
- implement interrupt synchronization (via an 10 ms  timer)
- don't make atusb_irq_enable call atusb_do_irq directly but start
  a tasklet instead. Otherwise, we recurse on a spin lock held
  somewhere in the interrupt handling infrastructure.

We send our first interrupt before at86rf230.c is ready to handle
it. This causes all further interrupts to be lost. As a work-around,
just clear IRQ_STATUS at the end of at86rf230_probe.

Signed-off-by: Werner Almesberger <werner@almesberger.net>