MIPS: JZ4740: A320: Debounce microswitch buttons.
On the microswitch buttons of the Dingoo A320 (L/R/select/start), it often
happens that multiple presses are detected for one actual button press.
Specifying a 5 ms debounce interval fixes that issue. I am not sure what
the optimum interval length would be, but 5 ms is short enough to not cause
problems for gamers and long enough to filter out the glitches.