MIPS: A320: Fine-tune PWM backlight settings.
The 1 ms PWM cycle is the same as legacy Dingux uses. We had some reports from
users for whom the 5 ms OD cycle was causing flickering. An added advantage is
that changing the backlight intensity via pwswd is more fluent now.

The default brightness was increased from 100 to 145, which is a better value
for use in the day time (not everyone is a night dweller like us ;).