MIPS: JZ4740: SLCD: In TV-out mode, add a black line on top and bottom
The bottom line ensures that the border color will be black instead of
whatever color the bottom-right pixel in the framebuffer has.

The top line delays the time that the descriptor for the main display
area is read, allowing vertical panning to be changed until just before
the display is sent to the TV encoder.

Paul Cercueil wrote the first version of this patch.