MIPS: JZ4740: Work around fbcon logo crash with SLCD.
The logo data is in the init section. However, when using the SLCD driver,
the fbcon code will try to fetch the logo data after the init section has
already been freed, potentially crashing the kernel.

This is a workaround because we have to release a stable kernel now.
The root cause should be found and fixed later.