MIPS: JZ4740: A320: Hardcode GPIO I2C bus number
The previous bus ID value of -1 should auto-assign an unused bus number,
but apparently it does not consider bus 0 unused since the RDA5807 is
registered on there, so it picks bus 1 and the RDA5807 becomes unreachable.

Thanks to Lars-Peter Clausen for figuring out the cause of this problem
and to Marcos Paulo de Souza for testing the fix.