media/radio/radio-rda5807.c: Wait for seek complete before return
All applications expects that the radio drivers make a synchronous
seek when trying to find radio stations. So this feature is
implemented here.

The rationale:
The hardware seek operation needs 35 msecs per frequency. We will
wait until the hardware finds a radio station, waiting 35 msecs for
each frequency (default spacing is 100 kHz).
When the flag STC (Seek/Tune Complete) is set, it means that the
hardware found a valid radio station, so we can return to the
application.

If we don't find a radio station, we will return -ETIMEDOUT to the
application.

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>

Maarten ter Huurne: Reworded comments, expressed timeout count using
                    constants rather than resulting value, return
                    immediately when tuned instead of breaking out
                    of the loop first.